+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * 【core   】     增加替换字符串中第一个指定字符串和最后一个指定字符串方法（pr#2533@Github）
 * 【jwt    】     JWT补充部分算法（pr#2546@Github）
 * 【core   】     NumberUtil.roundStr() 修改为使用toPlainString（pr#775@Gitee）
+* 【extra  】     QrCodeUtil新增SVG格式、Ascii Art字符画格式（pr#763@Gitee）
 * 
 ### 🐞Bug修复
 * 【http   】     修复https下可能的Patch、Get请求失效问题（issue#I3Z3DH@Gitee）